python/requirements.txt
requests
python-dotenv
pymysql
pytorch
numpy
pandas
matplotlib
bs4